backend/pyproject.toml
[project]
name = "ai-social-media-agent-backend"
version = "0.1.0"
description = "Backend for AI Social Media Agent"
authors = [
    {name = "yuchuchen"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.125.0,<0.126.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "langchain (>=1.2.0,<2.0.0)",
    "langgraph (>=1.0.5,<2.0.0)",
    "openai (>=2.14.0,<3.0.0)",
    "anthropic (>=0.75.0,<0.76.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "supabase (>=2.27.0,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "langchain-openai (>=1.1.6,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "pypdf (>=6.4.2,<7.0.0)",
    "vecs (>=0.4.5,<0.5.0)",
    "tavily-python (>=0.7.17,<0.8.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "langchain-google-genai (>=4.1.2,<5.0.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "chromadb (>=1.3.7,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
